+++ src/dipdup/config.py
@@ -907,6 +907,13 @@ class SentryConfig:
 
 
 @dataclass
+class PrometheusConfig:
+    host: str
+    port: int = 8000
+    update_interval: float = 1.0
+
+
+@dataclass
 class HookConfig(CallbackMixin, kind='hook'):
     args: Dict[str, str] = field(default_factory=dict)
     atomic: bool = False
@@ -989,6 +996,7 @@ class DipDupConfig:
     hooks: Dict[str, HookConfig] = field(default_factory=dict)
     hasura: Optional[HasuraConfig] = None
     sentry: Optional[SentryConfig] = None
+    prometheus: Optional[PrometheusConfig] = None
     advanced: AdvancedConfig = AdvancedConfig()
 
     def __post_init_post_parse__(self):